Refactored AuthService components
- Added new service: AuthService
- Moved GenerateJwtToken method from AuthController to AuthService
- Renamed method isCorrectPassword to IsCorrectPassword
- Moved IsCorrectPassword and HashPasswordForRegistration methods to AuthService

These changes improve adherence to the Single Responsibility Principle and naming conventions.
